pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wks-com"
description = "Communicate with your inverter (WKS EKO Circle)."
readme = "README.md"
requires-python = ">= 3.10"
authors = [
  { name = "Mickaël 'Tiger-222' Schoentgen", email="contact@tiger-222.fr" },
]
maintainers = [
  { name = "Mickaël 'Tiger-222' Schoentgen", email="contact@tiger-222.fr" },
]
license = { file = "LICENSE.txt" }
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: Information Technology",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: Unix",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
]
keywords = [
    "inverter",
    "rs232",
    "serial",
    "wks",
    "eko",
]
dependencies = [
  "pydantic>=2.5.2,<3",
  "pyserial>=3.5,<4",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/BoboTiG/python-wks-com"
Changelog = "https://github.com/BoboTiG/python-wks-com/blob/main/CHANGELOG.md"
Source = "https://github.com/BoboTiG/python-wks-com"
Sponsor = "https://github.com/sponsors/BoboTiG"
Tracker = "https://github.com/BoboTiG/python-wks-com/issues"
"Released Versions" = "https://github.com/BoboTiG/python-wks-com/releases"

[project.scripts]
wks-read = "wks_com.__main__:read"

[project.optional-dependencies]
dev = [
  "black",
  "build",
  "flake8-pyproject",
  "mypy",
  "isort",
  "pytest",
  "pytest-cov",
]

[tool.hatch.version]
path = "src/wks_com/__init__.py"

[tool.hatch.build]
skip-excluded-dirs = true

[tool.hatch.build.targets.sdist]
only-include = [
  "CHANGELOG.md",
  "src",
]

[tool.hatch.build.targets.wheel]
packages = [
  "src/wks_com",
]

[tool.black]
target-version = ["py310"]
line-length = 120
safe = true

[tool.flake8]
max-line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
# Ensure we know what we do
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true

# Imports management
ignore_missing_imports = true
follow_imports = "skip"

# Ensure full coverage
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true

# Restrict dynamic typing (a little)
# e.g. `x: List[Any]` or x: List`
disallow_any_generics = true

strict_equality = true

[tool.pytest.ini_options]
pythonpath = "src"
addopts = """
    --cov-report=term-missing:skip-covered
    --cov=src/wks_com
    --strict-markers
    -r fE
    -vvv
"""
